+++ CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## 0.5.11
+## 0.5.12
+- Add `dragent`: frontserver for DataRobot Agents
+- Add placeholder for E2E tests of `dragent` and `drmcp`
 
+## 0.5.11
 - Created DR MCP test stubs and added stub `client` attribute integrated with MCP server
 
 ## 0.5.10
-
 - Add DataRobotMCPServer APIs to retrieve registered tools/prompts/resources.
 
 ## 0.5.9